Create threadSafeSingleTon2.cpp
Optimised code for thread safe singleton  for multithreaded environment for better performance.
Makes use of the Double checking lock pattern logic.  
Only the initial access for threads that goes through the condition " if(instance == NULL)"  need to wait on lock.
Later on once object is created these threads never to need to obtain the lock and always uses the returned instance and gives better performance eventually.